fix: cleanup keyring
