+++ packages/gravity/kube-testing/package.json
@@ -21,7 +21,6 @@
     "@dxos/client": "workspace:*",
     "@dxos/client-services": "workspace:*",
     "@dxos/context": "workspace:*",
-    "@dxos/crypto": "workspace:*",
     "@dxos/debug": "workspace:*",
     "@dxos/echo-pipeline": "workspace:*",
     "@dxos/feed-store": "workspace:*",
