+++ packages/gravity/kube-testing/src/util.ts
@@ -2,14 +2,6 @@
 // Copyright 2023 DXOS.org
 //
 
-import { subtleCrypto } from '@dxos/crypto';
-
-export type TestKeyPair = {
-  privateKey: JsonWebKey;
-  publicKey: JsonWebKey;
-  publicKeyHex: string;
-};
-
 export const randomArraySlice = <T>(array: T[], size: number) => {
   const result = [];
   const arrayCopy = [...array];
@@ -20,27 +12,3 @@ export const randomArraySlice = <T>(array: T[], size: number) => {
   }
   return result;
 };
-
-export const generateKeyPair = async (): Promise<TestKeyPair> => {
-  const keyPair = await subtleCrypto.generateKey(
-    {
-      name: 'ECDSA',
-      namedCurve: 'P-256',
-    },
-    true,
-    ['sign', 'verify'],
-  );
-
-  const privateKeyExported = await subtleCrypto.exportKey('jwk', keyPair.privateKey);
-  const publicKeyExported = await subtleCrypto.exportKey('jwk', keyPair.publicKey);
-
-  // Convert the public key to hex format
-  const publicKeyBuffer = new Uint8Array(await subtleCrypto.exportKey('raw', keyPair.publicKey));
-  const publicKeyHex = Array.from(publicKeyBuffer).map(byte => byte.toString(16).padStart(2, '0')).join('');
-
-  return {
-    privateKey: privateKeyExported,
-    publicKey: publicKeyExported,
-    publicKeyHex,
-  };
-}
