+++ packages/gravity/kube-testing/tsconfig.json
@@ -25,9 +25,6 @@
       "path": "../../common/context"
     },
     {
-      "path": "../../common/crypto"
-    },
-    {
       "path": "../../common/debug"
     },
     {
